+++ src/features/index.ts
@@ -8,4 +8,4 @@ export * from './favorites/favorites.selectors';
 export * from './history/history.slice';
 export * from './history/history.actions';
 export * from './history/history.selectors';
-export * from './listener/listenerMiddleware';
+export * from './listener/listener';
